+++ mesonbuild/modules/qt4.py
