+++ mesonbuild/scripts/delwithsuffix.py
