+++ mesonconf
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from meson import mconf
+from mesonbuild import mconf
 import sys
 
 sys.exit(mconf.run(sys.argv[1:]))
