+++ mesongui
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from meson import mgui
+from mesonbuild import mgui
 import sys
 
 sys.exit(mgui.run(sys.argv))
